AuthService/AuthService/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
PyJWT==1.7.1
pytz==2022.2.1
sqlparse==0.4.2
